refactor(진성진): js 모듈 경로 수정 util -> common
