+++ src/js/signup.mjs
@@ -2,9 +2,9 @@ import {
   validate,
   resetErrorMessage,
   changeButtonStatus,
-} from "./utils/validate.mjs";
-import { toggleInputType } from "./utils/auth.mjs";
-import { authValidator } from "./constants/validator.mjs";
+} from "./common/validate.mjs";
+import { toggleInputType } from "./common/auth.mjs";
+import { authValidator } from "./common/validator.mjs";
 import { ERROR_MESSAGE } from "./constants/contants.mjs";
 
 const $emailInput = document.querySelector("#email");
@@ -21,7 +21,7 @@ const checkEmptyInputElements = [
   $repasswordInput,
 ];
 
-function checkFocusoutEmail(e) {
+function onFocusoutEmail(e) {
   resetErrorMessage(e);
   validate(e, {
     validator: authValidator.isEmptyInput,
@@ -83,7 +83,7 @@ function onClickButton(e) {
   location.href = "/signin.html";
 }
 
-$emailInput.addEventListener("focusout", checkFocusoutEmail);
+$emailInput.addEventListener("focusout", onFocusoutEmail);
 $nicknameInput.addEventListener("focusout", checkFocusoutNickname);
 $passwordInput.addEventListener("focusout", onFocusoutPassWord);
 $repasswordInput.addEventListener("focusout", onFocusoutRepassWord);
